Add the prstate action
This action updates who is assigned to a PR based on its state. If you
were requested for a review, you are automatically assigned. Likewise,
when you complete the review, you are unassigned. The same thing happens
with the submitter of the PR. When a PR has `CHANGES_REQUESTED`, the
submitter is assigned until they request rereview.