Rewrite prstate action
After thinking on this a bit, I want to move to a simpler responsibility
model. While there are any outstanding `CHANGES_REQUESTED`, the PR is
the responsiblity of the author. Otherwise, the PR is the responsibility
of the people who have been requested for review and the previous
reviewers who have had their reviews `DISMISSED`. The latter condition
captures reviewers who have accepted a PR, but then the author pushed
additional changes. In this case, the reviewer must re-review.

This responsibility model also has the nice property that reviewers
aren't responsible to re-review a PR until the author addresses all
outstanding reviews. Without this, reviews might become interleaved
which could result in conflicting requirements and a general increase of
disruption of engineer thought patterns.