Rewrite the prstate action
There were numerous problems with the previous version. We hopefully simplify
this by merging all events into a single stream and then processing them in
reverse order. This allows us to have later events modify how earlier events
are processed. The latest event that assigns responsibility, possibly in view
of later events, is the effective event.

This is a difficult problem and should probably be rewritten in GraphQL at some
point. But this is probably good enough for now.

Fixes: #25

Co-authored-by: Mark Bestavros <mbestavr@redhat.com>