Assign PRs to their authors on any CHANGES_REQUESTED reviews
Currently, a single CHANGES_REQUESTED review causes the ticket to be assigned
to the PR author and any additional requested reviewers who have not replied.
This has two negative consequences. First, there is no way to clearly place a
PR in the author's responsibility. For example: "Should we close this PR?"
Second, this encourages reviewers to submit reviews on top of ongoing changes
the author might make. This can only create frustration.

Therefore, this patch makes it ONLY the PR author's responsibility whenever a
CHANGES_REQUESTED review is submitted. No further reviews are needed until the
author addresses the current review.